requirements.txt
absl-py==1.4.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dm-tree==0.1.8
etils==1.8.0
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
importlib_resources==6.4.0
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.2.0
promise==2.3
protobuf==3.20.3
psutil==5.9.8
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-version==0.0.2
requests==2.31.0
rich==13.7.1
scipy==1.12.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-datasets==4.9.4
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
toml==0.10.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
XlsxWriter==1.1.2
xlwt==1.3.0
zipp==3.18.1
